Move everything to a jsonTransformer package; so that additional modules will not conflict package-wise.
