+++ json-transformer/src/main/java/com/expedia/www/haystack/pipes/jsonTransformer/ChangeEnvVarsToLowerCaseConfigurationSource.java
@@ -14,7 +14,7 @@
  *       limitations under the License.
  *
  */
-package com.expedia.www.haystack.pipes;
+package com.expedia.www.haystack.pipes.jsonTransformer;
 
 import org.cfg4j.source.ConfigurationSource;
 import org.cfg4j.source.context.environment.Environment;
