+++ json-transformer/src/main/java/com/expedia/www/haystack/pipes/jsonTransformer/Constants.java
@@ -14,8 +14,13 @@
  *       limitations under the License.
  *
  */
-package com.expedia.www.haystack.pipes;
+package com.expedia.www.haystack.pipes.jsonTransformer;
 
-public interface IntermediateStreamsConfig {
-    int replicationFactor();
+public interface Constants {
+    String SUBSYSTEM = "pipes";
+    String APPLICATION = "haystack-external-json-transformer";
+
+    // TODO Move topics to a centralized location to be used by all services
+    String KAFKA_FROM_TOPIC = "SpanObject-ProtobufFormat-Topic-1";
+    String KAFKA_TO_TOPIC = "SpanObject-JsonFormat-Topic-3";
 }
