+++ json-transformer/src/main/java/com/expedia/www/haystack/pipes/jsonTransformer/KafkaConfig.java
@@ -14,7 +14,7 @@
  *       limitations under the License.
  *
  */
-package com.expedia.www.haystack.pipes;
+package com.expedia.www.haystack.pipes.jsonTransformer;
 
 public interface KafkaConfig {
     String brokers();
