+++ json-transformer/src/main/java/com/expedia/www/haystack/pipes/jsonTransformer/SystemExitUncaughtExceptionHandler.java
@@ -14,7 +14,7 @@
  *       limitations under the License.
  *
  */
-package com.expedia.www.haystack.pipes;
+package com.expedia.www.haystack.pipes.jsonTransformer;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
